Add files via upload
This is Version 0.0 of a program I developed which auto-interprets petrophysical wireline logs from raw borehole data collated from conventional petroleum reservoirs. The program has such capabilities because the fundamental laws of conventional petrophysics combined with well-established petrophysical engineering rules of thumb are embedded within the code. This code enables - at the single click of a button (the execution of the code) - an immediate first-order understanding of the subsurface, without any manual input or interpretation from the user. 

The code writes interpreted well logs to an output .pdf and contains interpretations on fluids, rock competency, borehole rugosity, lithology, shale volume, and gas effect to name a few. 

This code would function excellently in a data room setting whereby two petroleum companies meet to discuss the purchase/sale of a particular field/reservoir. In such a scenario, usually only a limited amount of time is allowed to explore the data. This code yields a first interpretation within seconds and thus would be highly efficient in such a scenario. With this program, we can rapidly gain an understanding of whether an asset truly has potential to be valued at hundreds of millions of dollars or whether we are being sold a lie.  

The code is written in a simple manner to allow simple user manipulation. It requires input in the form of an Excel spreadsheet. However, a single line of code may be added first to convert a .las well log file into a .xslx, which can the be read by the program.

Code reference: Burtonshaw, J.E.J. (2021) A python program for automatic petrophysical interpretation. Version 0.0.